requirements.txt
Django
mypy
python-ldap
Celery
django-celery-results
django-celery-beat
flower
redis
hiredis
django-redis-cache
pytest-django
pytest-xdist
pytest-cov
psycopg2-binary
pytest-pgsql
